% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{active}
\alias{active}
\title{active}
\usage{
active(default = NULL, get = function() ..value.., set = function(value) value)
}
\arguments{
\item{default}{The default value of the object}

\item{get}{A function to get the value. It must take no arguments.  The value
returned by this function will be returned when the active binding is
called.}

\item{set}{A function to set the value. It must take one argument, the new
value.  The value returned will be assigned to the active binding.}
}
\value{
An object with class \code{active}
}
\description{
define a new active binding property
}
\details{
For both \code{get} and \code{set} functions, the value returned by the
function will be returned when the symbol is accessed or reassigned,
respectively.  You can use \code{self} to refer to the object within these
functions.  A special value of \code{..value..} will be accessible within these
functions; defaulting to the value in \code{default}.  The default value of
\code{get} uses this symbol.
}
\examples{
capsule <- new_capsule({
  x <- active(
    default = 0L,
    get = function() {
      ..value.. <<- ..value.. + 1L # increment value each time it's accessed
    },
    set = function(value) { # 'value' must be the only argument
      self@y <- value + 1L  # set another property on the object
      value                 # return object is assigned to the property
    }
  )

  y <- 0L
})


capsule@x
capsule@x
capsule@y
capsule@x <- 10L
capsule@x
capsule@y
}
