% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{barnard.check}
\alias{barnard.check}
\title{Verifying Barnard's Criterion}
\usage{
barnard.check(mat)
}
\arguments{
\item{mat}{array of dimensioins (N.T+1)x(N.C+1) where the (i,j) element is the order statistic for x.T=i and x.C=j}
}
\value{
logical. TRUE if Barnard's Criterion is satisfied, FALSE otherwise
}
\description{
This function checks whether Barnard's Criterion is verified
}
