% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{c.check}
\alias{c.check}
\title{Verifying Barnard's Criterion}
\usage{
\method{c}{check}(mat)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}
}
\value{
TRUE if Barnard's Criterion is satisfied, FALSE otherwise
}
\description{
This function checks whether Barnard's Criterion is verified
}
