% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.level}
\alias{chan.level}
\title{Level of Chan's Exact p-value}
\usage{
chan.level(alpha, N.T, N.C, Delta0, method)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
numeric representing the level using Chan's p-value method
}
\description{
This function calculates the level of the exact p-value based on Chan (1998)
}
