% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.mat}
\alias{chan.mat}
\title{Chan p-value for All 2x2 Tables}
\usage{
chan.mat(N.T, N.C, Delta0, method)
}
\arguments{
\item{N.T}{Sample size in Treatment group}

\item{N.C}{Sample size in Control group}

\item{Delta0}{Non-inferiority margin}

\item{method}{method for ordering criterion}
}
\value{
(N.T+1)x(N.C+1) array where the (i,j) element is the exact p-value for x.T=i and x.C=j
}
\description{
This function calculates the Chan p-values for all 2x2 tables given N.T, N.C and an ordering criterion
}
