% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{chan_level}
\alias{chan_level}
\title{Level of Chan's Exact p-value}
\usage{
chan_level(alpha, N.T, N.C, Delta0, method)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
numeric representing the level using Chan's p-value method
}
\description{
This function calculates the level of the exact p-value based on Chan (1998)
}
\examples{
#Level of Chan's p-value method for the Fries et al. study (1993)
chan_level(alpha=0.05,N.T=15,N.C=15,Delta0=0.2,method="MN")
}
\references{
\insertRef{Chan:03}{EC}
}
