% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{ci.level}
\alias{ci.level}
\title{Level of Confidence Interval}
\usage{
ci.level(alpha, N.T, N.C, Delta0, method, EC, tolEC, CZ, tolCZ, width)
}
\arguments{
\item{alpha}{significance level}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}

\item{EC}{TRUE if confidence limits are exact-corrected to match Chan's exact p-value, only relevant if CZ=F}

\item{tolEC}{tolerance for convergence}

\item{CZ}{TRUE if Chan and Zhang method used, FALSE otherwise}

\item{tolCZ}{increment size for domain of Delta, default set to 0.001}

\item{width}{range from starting values based on Miettinen & Nurminen confidence limits}
}
\value{
Level of test based on the specified confidence interval
}
\description{
This function evaluates true level of decision rule based on confidence interval method
}
