% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{ci_cz}
\alias{ci_cz}
\title{Confidence interval using Chan & Zhang (1999) methodology}
\usage{
ci_cz(x.T, x.C, N.T, N.C, method, alpha = 0.05, tol = 0.001, width = 0.3)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}

\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{tol}{positive numeric representing the increment size for domain of Delta. Default is set to 0.001.}

\item{width}{positive numeric representing the range from starting values based on Miettinen & Nurminen confidence limits.}
}
\value{
list of length 2 (D.lower, D.upper) representing the lower and upper Chan and Zhang confidence limits
}
\description{
This function calculates the confidence interval based on Chan & Zhang (1999)
}
\examples{
#Chan & Zhang confidence interval for the Rodary et al. study. (1989)
#ci_cz(x.T=83,x.C=69,N.T=88,N.C=76, method="MN")
}
\references{
\insertRef{Chan:99}{EC}
}
