% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{ci_ec}
\alias{ci_ec}
\title{Confidence interval using the exact-corrected methodology (Hawila & Berg, 2021)}
\usage{
ci_ec()
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}

\item{EC}{logical. TRUE for the exact-corrected confidence limits. FALSE for default method without exact-correction}

\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{tol}{positive numeric representing the tolerance for convergence}
}
\value{
list of length 2 (D.lower, D.upper) representing the lower and upper confidence limits
}
\description{
This is a wrapper function of \code{\link[EC]{confintZ}}
}
\examples{
#These two examples demonstrate the confidence intervals for the
#Rodary et al. study with and without the exact-correction.
confintZ(x.T=83,x.C=69,N.T=88,N.C=76,Delta0=0.1,method="MN", EC=TRUE)
confintZ(x.T=83,x.C=69,N.T=88,N.C=76,Delta0=0.1,method="MN", EC=FALSE)
}
\references{
\insertRef{Hawila:21}{EC}

\insertRef{Miettinen:85}{EC}

\insertRef{Farrington:90}{EC}
}
\seealso{
\link{confintZ}
}
