% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{ci_MN}
\alias{ci_MN}
\title{Confidence interval using the Miettinen & Nurminen (1985) methodology}
\usage{
ci_MN(x.T, x.C, N.T, N.C, alpha = 0.05, tol = 1e-10)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{tol}{positive numeric representing the tolerance for convergence}
}
\value{
list of length 2 (ci.lower, ci.upper) representing the lower and upper confidence limits
}
\description{
This is a wrapper function of \code{\link[EC]{confintZ}}
}
\examples{
#This examples demonstrates the confidence intervals for the
#Rodary et al. study using the exact-corrected method
ci_MN(x.T=83,x.C=69,N.T=88,N.C=76,alpha=0.05)
}
\references{
\insertRef{Miettinen:85}{EC}
}
\seealso{
\link{confintZ}
}
