% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walrus.R
\name{:=}
\alias{:=}
\title{Walrus operator}
\usage{
`:=`(sym, val)
}
\arguments{
\item{sym}{A variable name}

\item{val}{A value or expression with which \code{sym} is passed as a character,
unnamed, as the first argument to a function call, as a}
}
\description{
Walrus operator
}
