% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{cz_pval}
\alias{cz_pval}
\title{CZ p-value}
\usage{
cz_pval(x.T, x.C, N.T, N.C, Delta0, method, tol = 0.001)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}

\item{tol}{positive numeric representing the tolerance for calculation}
}
\value{
numeric representing the Chan & Zhang p-value
}
\description{
This function computes the Chan & Zhang p-value given in Hawila (2021)
}
\examples{
#This example is taken from Hawila (2021)
cz_pval(x.T=2,x.C=0,N.T=15,N.C=10,Delta0=0.12,method="MN")
}
\references{
\insertRef{Hawila:21}{EC}
}
