% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{.DollarNames.ec_capsule}
\alias{.DollarNames.ec_capsule}
\title{#' export
\verb{[[.ec_capsule} <- function(x, i, ...) {
if (i \%in\% names(x)) {
return(get(i, x, inherits = FALSE))
}}
\usage{
\method{.DollarNames}{ec_capsule}(x, pattern = "")
}
\description{

}
\details{
#' export
\verb{[[.ec_generator} <- function(x, i, ...) {
i <- substitute(i)
i <- as.character(i)
\verb{[[.ec_capsule}(environment(x), i, ...)
}

#' export
\verb{[.ec_capsule} <- function(x, i, ...) {
i <- substitute(i)
i <- as.character(i)
\verb{[[.ec_capsule}(x, i, ...)
}

#' export
\verb{[.ec_generator} <- function(x, i, ...) {
i <- substitute(i)
i <- as.character(i)
get(i, environment(x))
}

#' export
\verb{[[<-.ec_capsule} <- function(x, i, value) {
if (i \%in\% names(x)) {
# unlockBinding(i, x)
# on.exit(lockBinding(i, x), add = TRUE)
assign(i, value, x)
return(x)
}

stop(sprintf("No element named '\%s'", i))
}

#' export
\verb{[<-.ec_capsule} <- function(x, i, value) {
i <- substitute(i)
i <- as.character(i)
\verb{[[<-.ec_capsule}(x, i, value)
}
}
