% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-package.R
\docType{package}
\name{ec-package}
\alias{ec}
\alias{ec-package}
\title{ec: Encapsulated Classes}
\description{
Create simple, environment-based classes for object oriented programming.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
