% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-package.R
\name{Example}
\alias{Example}
\title{An example class}
\usage{
Example(x = 0L)
}
\arguments{
\item{x}{A non-negative integer to initialize the counter}
}
\description{
An example class
}
\section{Properties}{

\itemize{
\item \code{x}: An active binding representing the current value of the counter.
It must be a non-negative integer.
}

@section Methods:
\itemize{
\item \code{increment(n = 1L)}: Increment the counter by a non-negative integer \code{n}.
\item \code{reset()}: Reset the counter to zero.
}
}

