% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{likelihood.null}
\alias{likelihood.null}
\title{Null Likelihood}
\usage{
likelihood.null(x.T, x.C, N.T, N.C, P.T, Delta0)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{P.T}{proportion of responders in Treatment group}

\item{Delta0}{non-inferiority margin}
}
\value{
Exact p-value
}
\description{
This function evaluates the Binomial likelihood under P.T-P.C=-Delta0
}
