% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{likelihood_null}
\alias{likelihood_null}
\title{Null Likelihood}
\usage{
likelihood_null(x.T, x.C, N.T, N.C, P.T, Delta0)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{P.T}{numeric between 0 and 1 representing the proportion of responders in the treatment group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}
}
\value{
numeric representing the probability of getting the observed outcome under the null hypothesis
}
\description{
This function evaluates the Binomial likelihood under P.T-P.C=-Delta0
}
