% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{mat_Chan}
\alias{mat_Chan}
\title{Chan p-value for all 2x2 tables}
\usage{
mat_Chan(N.T, N.C, Delta0, method)
}
\arguments{
\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
array of dimension (N.T+1)x(N.C+1)  where the (i,j) element is the exact p-value for x.T=i and x.C=j
}
\description{
This function calculates the Chan p-values for all 2x2 tables given N.T, N.C and an ordering criterion;
This function is primarily used in the \code{\link[EC]{size_Chan}} function.
}
\examples{
#8x6 array of the the Chan p-values based on Farrington & Manning
#statistic for all possible 2x2 tables arising from a sample size
#of 7 in the treatment group and 5 in the control group and 10\%
#noninferiority margin.
mat_Chan(N.T=7,N.C=5,Delta0=0.1,method="FM")
}
\references{
\insertRef{Chan:98}{EC}
}
\seealso{
\link{size_Chan}
}
