% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule.R
\name{new_capsule}
\alias{new_capsule}
\title{Create a new class capsule}
\usage{
new_capsule(expr)
}
\arguments{
\item{expr}{An expression for assigning variables in new capsule.}
}
\value{
An environment with class \code{ec_capsule}
}
\description{
Create a new class capsule
}
\examples{
cap <- new_capsule({
  x <- 1
  y <- active(
    default = 2L,
    set = function(value) {
      self@x <- value
    },
    get = function() {
      ..value.. <<- ..value.. + 1L
      ..value..
    }
  )

  reset <- function() {
    self@y <- 0L
    invisible(self)
  }
})
cap@y
cap@y
}
