% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{order_mat}
\alias{order_mat}
\title{Order statistic for all 2x2 tables}
\usage{
order_mat(N.T, N.C, Delta0, method)
}
\arguments{
\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
array of dimensions (N.T+1)x(N.C+1) where the (i,j) element is the order statistic for x.T=i and x.C=j
}
\description{
This function calculate the order statistics for all 2x2 tables
}
\examples{
#16x16 array of the the Santner & Snell test statistic for all
#possible 2x2 tables arising from a sample size of 15 in the
#treatment and control group and 30\% noninferiority margin.
order_mat(N.T=15,N.C=15,Delta0=0.3,method="SS")
}
\references{
\insertRef{Santner:80}{EC}
}
