% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{orderfun}
\alias{orderfun}
\title{Test Statistic for Non-inferiority Hypothesis}
\usage{
orderfun(x.T, x.C, N.T, N.C, Delta0, method)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}
}
\value{
statistic value of ordering criterion
}
\description{
This function  calculate statistics for testing non-inferiority based on (Santner & Snell (1980), Blackwelder (1982),
Miettinen & Nurminen (1985) and Farrington & Manning (1990))
}
