% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{orderfunEC}
\alias{orderfunEC}
\title{Exact-Corrected Test Statistic}
\usage{
orderfunEC(x.T, x.C, N.T, N.C, Delta, Delta0)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta}{constraint for P.C-P.T}

\item{Delta0}{non-inferiority margin}
}
\value{
Exact-Corrected statistic
}
\description{
This function calculates the Exact-Corrected test statistic
}
