% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{pval_Wald}
\alias{pval_Wald}
\title{p-value corresponding to the Wald confidence interval}
\usage{
pval_Wald(x.T, x.C, N.T, N.C, delta0)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{delta0}{numeric between 0 and 1 representing the noninferiority margin}
}
\value{
numeric representing the asymptotic p-value associated with the Miettinen & Nurminen statistic
}
\description{
This function calculates the Wald p-value
}
\examples{
#The example is taken from Rodary et al. (1989) which was analyzed by Chan (1998)
pval_MN(x.T=83,x.C=69,N.T=88,N.C=76,delta0=0.1)
}
