% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{restricted_mle}
\alias{restricted_mle}
\title{Restricted MLE of variance}
\usage{
restricted_mle(x.T, x.C, N.T, N.C, Delta)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta}{numeric between -1 and 1 representing the constraint for P.C-P.T}
}
\value{
list of length 2 where the first element is MR.T representing the MLE for the treatment group and the second element is MR.C representing the MLE for the control group
}
\description{
This function calculates the maximum likelihood estimator (MLE) of the variance
restricted to P.T-P.C=-Delta.
}
\examples{
#calculates MLE of variance of 5/10 and 7/10 under the constraint P.T=P.C-0.2
restricted_mle(x.T=5,x.C=7,N.T=10,N.C=10,Delta=0.2)
}
\references{
\insertRef{Miettinen:85}{EC}
}
