% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{size_Chan}
\alias{size_Chan}
\title{Maximal size of Chan's exact p-value}
\usage{
size_Chan(alpha, N.T, N.C, Delta0, method)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
numeric representing the level using Chan's p-value method
}
\description{
This function calculates the maximal size of Chan's (1998) exact p-value; see Hawila & Berg (2021)
}
\examples{
#Level of Chan's p-value method for the Fries et al. study (1993)
size_Chan(alpha=0.05,N.T=15,N.C=15,Delta0=0.2,method="MN")
}
\references{
\insertRef{Hawila:21}{EC}
\insertRef{Chan:03}{EC}
}
