% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{size_CZ}
\alias{size_CZ}
\title{Maximal size of the Chan & Zhang (1999) confidence interval}
\usage{
size_CZ(alpha, N.T, N.C, delta0, tolCZ = 0.001, eps = 0.001)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{tolCZ}{positive numeric representing the increment size for domain of Delta. Default is set to 0.001.}

\item{eps}{positive numeric representing the range from starting values based on Miettinen & Nurminen confidence limits.}
}
\value{
numeric representing the level of test based on the specified confidence interval method
}
\description{
This function evaluates maximum size of the decision rule based on Chan & Zhang's confidence interval
}
\examples{
#The three examples calculate the level of the exact-corrected,
#Chan & Zhang and Asymptotic confidence interval based on the
#Miettenen and Nurminen test statistic where each group has a
#sample size of 10, alpha is 0.1 and the noninferiority margin
#is 20\%.
size_CZ(alpha=0.1,N.T=10,N.C=10,delta0=0.2,tolCZ=1e-3)
}
\references{
\insertRef{Hawila:21}{EC}
}
