% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{size_EC}
\alias{size_EC}
\title{Maximal size of the exact-corrected confidence interval}
\usage{
size_EC(alpha, N.T, N.C, delta0, tolEC = 1e-04)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{tolEC}{positive numeric representing the tolerance for confidence interval convergence}
}
\value{
numeric representing the level of test based on the specified confidence interval method
}
\description{
This function evaluates maximum size of the decision rule based on the exact-corrected confidence interval
}
\examples{
#The three examples calculate the level of the exact-corrected,
#Chan & Zhang and Asymptotic confidence interval based on the
#Miettenen and Nurminen test statistic where each group has a
#sample size of 10, alpha is 0.1 and the noninferiority margin
#is 20\%.
size_EC(alpha=0.1,N.T=10,N.C=10,delta0=0.2,tolEC=1e-4)
}
\references{
\insertRef{Hawila:21}{EC}
}
