% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{size_general}
\alias{size_general}
\title{Maximal size of tests based on various confidence interval methods}
\usage{
size_general(
  alpha,
  N.T,
  N.C,
  delta0,
  method = "MN",
  EC,
  tolEC = 1e-06,
  CZ,
  tolCZ = 0.001,
  eps = 0.3
)
}
\arguments{
\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}

\item{EC}{logical. TRUE for the exact-corrected confidence limits. FALSE for default method without exact-correction. Only relevant if CZ=F}

\item{tolEC}{positive numeric representing the tolerance for confidence interval convergence}

\item{CZ}{logical. TRUE for Chan and Zhang confidence limits. FALSE for Exact-corrected (EC=T) or default method without correction (EC=F)}

\item{tolCZ}{positive numeric representing the increment size for domain of Delta. Default is set to 0.001.}

\item{eps}{positive numeric representing the range from starting values based on Miettinen & Nurminen confidence limits.}
}
\value{
numeric representing the level of test based on the specified confidence interval method
}
\description{
This function evaluates maximum size of the decision rule based on a given confidence interval method; see Hawila & Berg (2021)
}
\examples{
#The three examples calculate the level of the exact-corrected,
#Chan & Zhang and Asymptotic confidence interval based on the
#Miettenen and Nurminen test statistic where each group has a
#sample size of 10, alpha is 0.1 and the noninferiority margin
#is 20\%.
size_general(alpha=0.1,N.T=10,N.C=10,delta0=0.2,method="MN",EC=TRUE,tolEC=1e-4,CZ=FALSE,tolCZ=1e-3,eps=0.3)
#size_general(alpha=0.1,N.T=10,N.C=10,delta0=0.2,method="MN",EC=FALSE,tolEC=1e-4,CZ=TRUE,tolCZ=1e-3,eps=0.3)
size_general(alpha=0.1,N.T=10,N.C=10,delta0=0.2,method="MN",EC=FALSE,tolEC=1e-4,CZ=FALSE,tolCZ=1e-3,eps=0.3)
}
\references{
\insertRef{Hawila:21}{EC}
\insertRef{Chan:99}{EC}
\insertRef{Miettinen:85}{EC}
}
