% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{Barnard_check}
\alias{Barnard_check}
\title{Verifying Barnard's criterion}
\usage{
Barnard_check(mat)
}
\arguments{
\item{mat}{array of dimensions (N.T+1)x(N.C+1) where the (i,j) element is the order statistic for x.T=i and x.C=j}
}
\value{
logical. TRUE if Barnard's criterion is satisfied, FALSE otherwise
}
\description{
This function checks whether Barnard's criterion (Biometrika, 1947) is satisfied
}
\examples{
#Checks if the Barnard criterion is satisfied for the
#Miettenin & Nurminen ordering statistic with N.T=15,
#N.C=15 and delta0=30\%
Barnard_check(order_mat(N.T=15,N.C=15,delta0=0.3,method="MN"))
}
\references{
\insertRef{Barnard:47}{EC}
}
