% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.level}
\alias{chan.level}
\title{Level of Chan's Exact p-value}
\usage{
chan.level(alpha, N.T, N.C, Delta0, method)
}
\arguments{
\item{alpha}{significance level}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}
}
\value{
True level of Chan's p-value
}
\description{
This function calculates the level of the exact p-value based on Chan (1998)
}
