% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.mat}
\alias{chan.mat}
\title{Chan p-value for All 2x2 Tables}
\usage{
chan.mat(N.T, N.C, Delta0, method)
}
\arguments{
\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}

\item{method}{character representing the method for ordering criterion("MN","FM","SS","Blackwelder")}
}
\value{
array of dimension (N.T+1)x(N.C+1)  where the (i,j) element is the exact p-value for x.T=i and x.C=j
}
\description{
This function calculates the Chan p-values for all 2x2 tables given N.T, N.C and an ordering criterion
}
