% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{ci_EC}
\alias{ci_EC}
\title{Confidence interval using the exact-corrected methodology (Hawila & Berg, 2021)}
\usage{
ci_EC(x.T, x.C, N.T, N.C, Delta0, alpha = 0.05, tol = 1e-10)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{Delta0}{numeric between 0 and 1 representing the noninferiority margin}

\item{alpha}{numeric between 0 and 1 representing the significance level}

\item{tol}{positive numeric representing the tolerance for convergence}
}
\value{
list of length 2 (ci.lower, ci.upper) representing the lower and upper confidence limits
}
\description{
This is a wrapper function of \code{\link[EC]{ci_general}}
}
\examples{
#This examples demonstrates the confidence intervals for the
#Rodary et al. study using the exact-corrected method
ci_EC(x.T=83,x.C=69,N.T=88,N.C=76,Delta0=0.1,alpha=0.05)
}
\references{
\insertRef{Hawila:21}{EC}
}
\seealso{
\link{ci_general}
}
