% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{likelihood_null}
\alias{likelihood_null}
\title{Null Likelihood}
\usage{
likelihood_null(x.T, x.C, N.T, N.C, P.T, Delta0)
}
\arguments{
\item{x.T}{positive integer representing the observed number of responders in the treatment group}

\item{x.C}{positive integer representing the observed number of responders in the control group}

\item{N.T}{positive integer representing the sample size in the treatment group}

\item{N.C}{positive integer representing the sample size in the control group}

\item{P.T}{numeric between 0 and 1 representing the proportion of responders in the treatment group}

\item{Delta0}{numeric between 0 and 1 representing the non-inferiority margin}
}
\value{
numeric representing the probability of getting the observed outcome under the null hypothesis
}
\description{
This function evaluates the Binomial likelihood under P.T-P.C=-Delta0
}
\examples{
likelihood_null(x.T=10, x.C=8, N.T=20, N.C=20, P.T=0.3, Delta0=0.1)
#The probability of getting 10/20 successes in the treatment group and 8/20 in the placebo group when the proportion of responders is 30\% and the noninferiority margin is 10\% can be calculated by
}
